+++ netbanking/src/androidTest/java/cz/csas/netbanking/accounts/AccountsTransactionsHistoryListPage0Test.java
@@ -28,7 +28,7 @@ import static junit.framework.Assert.assertFalse;
 public class AccountsTransactionsHistoryListPage0Test extends NetbankingTest {
     private final String X_JUDGE_CASE = "accounts.withId.transactionsHistory.list.page0";
     private CountDownLatch mAccountsTransactionsHistorySignal;
-    private AccountTransactionsListResponse mAccountTransactionsListResponse;
+    private AccountTransactionsHistoryListResponse mAccountTransactionsHistoryListResponse;
 
     @Override
     public void setUp() {
@@ -43,16 +43,16 @@ public class AccountsTransactionsHistoryListPage0Test extends NetbankingTest {
     @Test
     public void testAccountTransactionsHistoryListPage0() {
 
-        AccountTransactionsParameters parameters = new AccountTransactionsParameters.Builder()
+        AccountTransactionsHistoryParameters parameters = new AccountTransactionsHistoryParameters.Builder()
                 .setDateStart(TimeUtils.getISO8601Date("2014-06-01T00:00:00+02:00"))
                 .setDateEnd(TimeUtils.getISO8601Date("2014-06-30T00:00:00+02:00"))
                 .setPagination(new Pagination(0, 1))
                 .build();
 
-        mNetbankingClient.getAccountsResource().withId("CZ5508000000000379554193").getTransactionsHistoryResource().list(parameters, new CallbackWebApi<AccountTransactionsListResponse>() {
+        mNetbankingClient.getAccountsResource().withId("CZ5508000000000379554193").getTransactionsHistoryResource().list(parameters, new CallbackWebApi<AccountTransactionsHistoryListResponse>() {
             @Override
-            public void success(AccountTransactionsListResponse accountTransactionsListResponse) {
-                mAccountTransactionsListResponse = accountTransactionsListResponse;
+            public void success(AccountTransactionsHistoryListResponse accountTransactionsHistoryListResponse) {
+                mAccountTransactionsHistoryListResponse = accountTransactionsHistoryListResponse;
                 mAccountsTransactionsHistorySignal.countDown();
             }
 
@@ -68,7 +68,7 @@ public class AccountsTransactionsHistoryListPage0Test extends NetbankingTest {
             e.printStackTrace();
         }
 
-        AccountTransaction transaction = mAccountTransactionsListResponse.getAccountTransactions().get(0);
+        AccountTransaction transaction = mAccountTransactionsHistoryListResponse.getAccountTransactions().get(0);
         AccountParty accountParty = transaction.getAccountParty();
         Amount amount = transaction.getAmount();
         Amount amountSender = transaction.getAmountSender();